# deep-research-orchestrator/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Deep Research Orchestrator API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/settings/llm-profiles": {
   "get": {
    "tags": [
     "settings"
    ],
    "summary": "List Profiles",
    "operationId": "list_profiles_api_settings_llm_profiles_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/LlmProfileView"
         },
         "type": "array",
         "title": "Response List Profiles Api Settings Llm Profiles Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "settings"
    ],
    "summary": "Create Profile",
    "operationId": "create_profile_api_settings_llm_profiles_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LlmProfileIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmProfileView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/settings/llm-profiles/{profile_id}": {
   "put": {
    "tags": [
     "settings"
    ],
    "summary": "Update Profile",
    "operationId": "update_profile_api_settings_llm_profiles__profile_id__put",
    "parameters": [
     {
      "name": "profile_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Profile Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LlmProfileIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmProfileView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "settings"
    ],
    "summary": "Delete Profile",
    "operationId": "delete_profile_api_settings_llm_profiles__profile_id__delete",
    "parameters": [
     {
      "name": "profile_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Profile Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/settings/llm-profiles/{profile_id}/test": {
   "post": {
    "tags": [
     "settings"
    ],
    "summary": "Test Profile",
    "description": "接続試験: 到達性・認証・model有無・最小生成。\n\n注意: 有料provider (openai/anthropic) では最小request分の実課金が発生し得る。\nUIは実行前にその旨を表示する。",
    "operationId": "test_profile_api_settings_llm_profiles__profile_id__test_post",
    "parameters": [
     {
      "name": "profile_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Profile Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Test Profile Api Settings Llm Profiles  Profile Id  Test Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/settings/roles": {
   "get": {
    "tags": [
     "settings"
    ],
    "summary": "Get Roles",
    "operationId": "get_roles_api_settings_roles_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "anyOf": [
           {
            "type": "string"
           },
           {
            "type": "null"
           }
          ]
         },
         "type": "object",
         "title": "Response Get Roles Api Settings Roles Get"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "settings"
    ],
    "summary": "Put Roles",
    "operationId": "put_roles_api_settings_roles_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleAssignmentIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "anyOf": [
           {
            "type": "string"
           },
           {
            "type": "null"
           }
          ]
         },
         "type": "object",
         "title": "Response Put Roles Api Settings Roles Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/settings/proxy": {
   "get": {
    "tags": [
     "settings"
    ],
    "summary": "Get Proxy",
    "operationId": "get_proxy_api_settings_proxy_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ProxyConfigView"
         },
         "type": "array",
         "title": "Response Get Proxy Api Settings Proxy Get"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "settings"
    ],
    "summary": "Put Proxy",
    "operationId": "put_proxy_api_settings_proxy_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProxyConfigIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProxyConfigView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/settings/proxy/test": {
   "post": {
    "tags": [
     "settings"
    ],
    "summary": "Test Proxy",
    "description": "Test proxy: 外部URLがproxy経由になり、internal/Local LLMがbypassされることを確認。",
    "operationId": "test_proxy_api_settings_proxy_test_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProxyTestIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Test Proxy Api Settings Proxy Test Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/settings/llm-endpoint-allowlist/{entry_id}": {
   "delete": {
    "tags": [
     "settings"
    ],
    "summary": "Delete Allowlist Entry",
    "operationId": "delete_allowlist_entry_api_settings_llm_endpoint_allowlist__entry_id__delete",
    "parameters": [
     {
      "name": "entry_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entry Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LlmProfileIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Name"
     },
     "provider": {
      "type": "string",
      "enum": [
       "local",
       "openai",
       "anthropic"
      ],
      "title": "Provider"
     },
     "api": {
      "type": "string",
      "enum": [
       "openai-compatible",
       "anthropic"
      ],
      "title": "Api",
      "default": "openai-compatible"
     },
     "endpoint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Endpoint"
     },
     "model": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Model"
     },
     "api_key": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Key"
     },
     "timeout_seconds": {
      "type": "integer",
      "maximum": 3600.0,
      "minimum": 5.0,
      "title": "Timeout Seconds",
      "default": 120
     },
     "max_concurrency": {
      "type": "integer",
      "maximum": 32.0,
      "minimum": 1.0,
      "title": "Max Concurrency",
      "default": 2
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "name",
     "provider",
     "model"
    ],
    "title": "LlmProfileIn"
   },
   "LlmProfileView": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "provider": {
      "type": "string",
      "title": "Provider"
     },
     "api": {
      "type": "string",
      "title": "Api"
     },
     "endpoint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Endpoint"
     },
     "model": {
      "type": "string",
      "title": "Model"
     },
     "has_api_key": {
      "type": "boolean",
      "title": "Has Api Key"
     },
     "api_key_masked": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Key Masked"
     },
     "timeout_seconds": {
      "type": "integer",
      "title": "Timeout Seconds"
     },
     "max_concurrency": {
      "type": "integer",
      "title": "Max Concurrency"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "provider",
     "api",
     "endpoint",
     "model",
     "has_api_key",
     "api_key_masked",
     "timeout_seconds",
     "max_concurrency",
     "enabled"
    ],
    "title": "LlmProfileView"
   },
...
   "ProxyConfigIn": {
    "properties": {
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": "global"
     },
     "mode": {
      "type": "string",
      "enum": [
       "off",
       "inherit",
       "explicit"
      ],
      "title": "Mode"
     },
     "http_proxy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Http Proxy"
     },
     "https_proxy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Https Proxy"
     },
     "all_proxy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "All Proxy"
     },
     "no_proxy": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "No Proxy"
     },
     "ca_bundle_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ca Bundle Path"
     }
    },
    "type": "object",
    "required": [
     "mode"
    ],
    "title": "ProxyConfigIn"
   },
   "ProxyConfigView": {
    "properties": {
     "scope": {
      "type": "string",
      "title": "Scope"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "has_http_proxy": {
      "type": "boolean",
      "title": "Has Http Proxy"
     },
     "has_https_proxy": {
      "type": "boolean",
      "title": "Has Https Proxy"
     },
     "has_all_proxy": {
      "type": "boolean",
      "title": "Has All Proxy"
     },
     "no_proxy": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "No Proxy"
     },
     "ca_bundle_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ca Bundle Path"
     }
    },
    "type": "object",
    "required": [
     "scope",
     "mode",
     "has_http_proxy",
     "has_https_proxy",
     "has_all_proxy",
     "no_proxy",
     "ca_bundle_path"
    ],
    "title": "ProxyConfigView"
   },
   "ProxyTestIn": {
    "properties": {
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": "global"
     },
     "external_url": {
      "type": "string",
      "title": "External Url",
      "default": "https://example.com/"
     },
     "internal_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Internal Url"
     }
    },
    "type": "object",
    "title": "ProxyTestIn"
   },
   "RoleAssignmentIn": {
    "properties": {
     "assignments": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "object",
      "title": "Assignments"
     }
    },
    "type": "object",
    "required": [
     "assignments"
    ],
    "title": "RoleAssignmentIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}